feat: intoduce `aptos-verifier`
