+++ lib/aptos-verifier/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "aptos-verifier"
+version = "0.1.0"
+edition.workspace = true
+license-file.workspace = true
+repository.workspace = true
+
+[dependencies]
+unionlabs = { workspace = true }
+thiserror = { workspace = true }
+sha3 = { workspace = true }
+
+[lints]
+workspace = true
