+++ light-clients/movement/ics08-movement/src/lib.rs
@@ -2,7 +2,6 @@ mod client;
 mod contract;
 mod errors;
 mod sparse_merkle_proof;
-mod transaction_proof;
 mod types;
 
 // use aptos_types::{proof::TransactionInfoWithProof, state_proof::StateProof};
